backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
pyjwt==2.8.0
ultralytics==8.2.0
opencv-python-headless
pyserial==3.5
numpy
gunicorn
python-dotenv
google-generativeai
